voice-agent/pyproject.toml
[project]
name = "voice-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents[google]",
    "livekit-plugins-noise-cancellation>=0.2.5",
    "python-dotenv>=1.2.1",
]
